requirements.txt
requests==2.32.3
python-dotenv==1.0.1
opencv-python==4.10.0.84
pydantic==2.9.2
paho-mqtt==2.1.0
schedule==1.2.2
ultralytics==8.2.78